+++ misskey-api/src/endpoint/channels/update.rs
@@ -21,11 +21,22 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub banner_id: Option<Option<Id<DriveFile>>>,
+    #[cfg(feature = "13-12-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-12-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub is_archived: Option<bool>,
     #[cfg(feature = "13-11-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub pinned_note_ids: Option<Vec<Id<Note>>>,
+    /// [ 1 .. 16 ] characters
+    #[cfg(feature = "13-12-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-12-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub color: Option<String>,
 }
 
 impl misskey_core::Request for Request {
@@ -55,8 +66,12 @@ mod tests {
             name: Some("yCdKKHkRAmqhE49j3TBCVnnsQiZ2KkCK83z6NNKoGaiqQNOALsAOIz6XVJAcaV9lNbQYuuhSe7nAM8qdvUtokhWYWDO999z07N4ajtikDmzANpgwRh7rxMgb8PLsgcbm".to_string()),
             description: None,
             banner_id: None,
+            #[cfg(feature = "13-12-0")]
+            is_archived: None,
             #[cfg(feature = "13-11-0")]
             pinned_note_ids: None,
+            #[cfg(feature = "13-12-0")]
+            color: None
         }).await;
     }
 
@@ -77,8 +92,12 @@ mod tests {
                 name: None,
                 description: Some(Some("description".to_string())),
                 banner_id: None,
+                #[cfg(feature = "13-12-0")]
+                is_archived: None,
                 #[cfg(feature = "13-11-0")]
                 pinned_note_ids: None,
+                #[cfg(feature = "13-12-0")]
+                color: None,
             })
             .await;
         client
@@ -87,8 +106,37 @@ mod tests {
                 name: None,
                 description: Some(None),
                 banner_id: None,
+                #[cfg(feature = "13-12-0")]
+                is_archived: None,
                 #[cfg(feature = "13-11-0")]
                 pinned_note_ids: None,
+                #[cfg(feature = "13-12-0")]
+                color: None,
+            })
+            .await;
+    }
+
+    #[cfg(feature = "13-12-0")]
+    #[tokio::test]
+    async fn request_with_is_archived() {
+        let client = TestClient::new();
+        let channel = client
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test channel")
+                    .build(),
+            )
+            .await;
+
+        client
+            .test(Request {
+                channel_id: channel.id.clone(),
+                name: None,
+                description: None,
+                banner_id: None,
+                is_archived: Some(true),
+                pinned_note_ids: None,
+                color: None,
             })
             .await;
     }
@@ -112,8 +160,12 @@ mod tests {
                 name: None,
                 description: None,
                 banner_id: Some(Some(file.id)),
+                #[cfg(feature = "13-12-0")]
+                is_archived: None,
                 #[cfg(feature = "13-11-0")]
                 pinned_note_ids: None,
+                #[cfg(feature = "13-12-0")]
+                color: None,
             })
             .await;
         // bug in misskey
@@ -132,11 +184,11 @@ mod tests {
     async fn request_with_pinned_note_ids() {
         let client = TestClient::new();
         let channel = client
-            .test(crate::endpoint::channels::create::Request {
-                name: "test channel".to_string(),
-                description: None,
-                banner_id: None,
-            })
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test channel")
+                    .build(),
+            )
             .await;
         let note = client.create_note(Some("test"), None, None).await;
 
@@ -146,7 +198,36 @@ mod tests {
                 name: None,
                 description: None,
                 banner_id: None,
+                #[cfg(feature = "13-12-0")]
+                is_archived: None,
                 pinned_note_ids: Some(vec![note.id]),
+                #[cfg(feature = "13-12-0")]
+                color: None,
+            })
+            .await;
+    }
+
+    #[cfg(feature = "13-12-0")]
+    #[tokio::test]
+    async fn request_with_color() {
+        let client = TestClient::new();
+        let channel = client
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test channel")
+                    .build(),
+            )
+            .await;
+
+        client
+            .test(Request {
+                channel_id: channel.id.clone(),
+                name: None,
+                description: None,
+                banner_id: None,
+                is_archived: None,
+                pinned_note_ids: None,
+                color: Some("#ff0000".to_string()),
             })
             .await;
     }