+++ misskey-util/src/builder/channel.rs
@@ -19,6 +19,8 @@ impl<C> ChannelBuilder<C> {
             name: String::default(),
             description: None,
             banner_id: None,
+            #[cfg(feature = "13-12-0")]
+            color: None,
         };
         ChannelBuilder { client, request }
     }
@@ -43,6 +45,14 @@ impl<C> ChannelBuilder<C> {
     /// Sets the banner image of the channel.
     pub fn banner(&mut self, file: impl EntityRef<DriveFile>) -> &mut Self {
         self.request.banner_id.replace(file.entity_ref());
+        self
+    }
+
+    /// Sets the color of the channel.
+    #[cfg(feature = "13-12-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-12-0")))]
+    pub fn color(&mut self, color: impl Into<String>) -> &mut Self {
+        self.request.color.replace(color.into());
         self
     }
 }
@@ -75,8 +85,12 @@ impl<C> ChannelUpdateBuilder<C> {
             name: None,
             description: None,
             banner_id: None,
+            #[cfg(feature = "13-12-0")]
+            is_archived: None,
             #[cfg(feature = "13-11-0")]
             pinned_note_ids: None,
+            #[cfg(feature = "13-12-0")]
+            color: None,
         };
         ChannelUpdateBuilder { client, request }
     }
@@ -101,6 +115,14 @@ impl<C> ChannelUpdateBuilder<C> {
         pub banner: impl EntityRef<DriveFile> { banner_id = banner.entity_ref() };
     }
 
+    /// Sets whether the channel is archived.
+    #[cfg(feature = "13-12-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-12-0")))]
+    pub fn archived(&mut self, archived: bool) -> &mut Self {
+        self.request.is_archived.replace(archived);
+        self
+    }
+
     /// Sets the pinned notes of the channel.
     #[cfg(feature = "13-11-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
@@ -111,6 +133,14 @@ impl<C> ChannelUpdateBuilder<C> {
         self.request
             .pinned_note_ids
             .replace(notes.into_iter().map(|note| note.entity_ref()).collect());
+        self
+    }
+
+    /// Sets the color of the channel.
+    #[cfg(feature = "13-12-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-12-0")))]
+    pub fn color(&mut self, color: impl Into<String>) -> &mut Self {
+        self.request.color.replace(color.into());
         self
     }
 }