+++ misskey-api/CHANGELOG.md
@@ -115,6 +115,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Misskey v13.11.3
    - endpoint `roles/note`
    - channel `roleTimeline`
+- Support for Misskey v13.12.0
+   - endpoint `admin/emoji/set-license-bulk`
+   - endpoint `users/update-memo`
 
 ### Changed